Allow substrateFun reading less, but not more
Partial BER decoding was possible in 0.4.8 and is no longer possible in
0.5.0. It's currently prevented by the strict bytesRead check in state
stDecodeValue. We have to relax length check a bit.

As a custom substrateFuns can be used for partial decoding, reading less
than the definite length is fine and expected. Reading more is fishy.
Treat the latter as error. If the length check breaks your existing
code, please file a bug for pyasn1 and explain the use case for reading
too much.