Fix link to "Layman's Guide"
The host ftp.rsasecurity.com does no longer resolve for me.

kk:~ kris$ host ftp.rsasecurity.com
Host ftp.rsasecurity.com not found: 3(NXDOMAIN)

FTP Links are no longer supported directly in many browsers.

I replace the link to the text to a PDF version of the document in question, hosted by researchgate.net. This is supposed to be a stable URL, and the formatted document is probably easier to read.